+++ packages/core/src/utils/make-namespaced.ts
@@ -17,10 +17,9 @@ const OUTPUT_SELECTION = {
  *
  * This makes the following modifications to the input:
  * - Adds a state variable for each namespaced struct definition
- * - Deletes all contracts' functions since they are not needed for storage layout
- * - Deletes all contracts' modifiers, variables, and parent constructor invocations to avoid compilation errors due to deleted functions and constructors
- * - Deletes all using for directives (at file level and in contracts) since they may reference deleted functions
- * - Converts all free functions and constants (at file level) to dummy variables (do not delete them since they might be imported by other files)
+ * - For each contract, for all node types that are not needed for storage layout or may reference deleted functions and constructors, converts them to dummy enums with random id
+ * - Converts all using for directives (at file level and in contracts) to dummy enums with random id (do not delete them to avoid orphaning possible NatSpec documentation)
+ * - Converts all custom errors, free functions and constants (at file level) to dummy enums with the same name (do not delete them since they might be imported by other files)
  *
  * Also sets the outputSelection to only include storageLayout and ast, since the other outputs are not needed.
  *
@@ -41,55 +40,95 @@ export function makeNamespacedInput(input: SolcInput, output: SolcOutput): SolcI
 
     const orig = Buffer.from(source.content, 'utf8');
 
+    const replacedIdentifiers = new Set<string>();
     const modifications: Modification[] = [];
 
     for (const node of output.sources[sourcePath].ast.nodes) {
-      if (isNodeType('ContractDefinition', node)) {
-        const contractDef = node;
+      switch (node.nodeType) {
+        case 'ContractDefinition': {
+          const contractDef = node;
 
-        // Remove any calls to parent constructors from the inheritance list
-        const inherits = contractDef.baseContracts;
-        for (const inherit of inherits) {
-          if (Array.isArray(inherit.arguments)) {
-            assert(inherit.baseName.name !== undefined);
-            modifications.push(makeReplace(inherit, orig, inherit.baseName.name));
-          }
-        }
-
-        const contractNodes = contractDef.nodes;
-        for (const contractNode of contractNodes) {
-          if (
-            isNodeType('FunctionDefinition', contractNode) ||
-            isNodeType('ModifierDefinition', contractNode) ||
-            isNodeType('VariableDeclaration', contractNode)
-          ) {
-            if (contractNode.documentation) {
-              modifications.push(makeDelete(contractNode.documentation, orig));
-            }
-            modifications.push(makeDelete(contractNode, orig));
-          } else if (isNodeType('UsingForDirective', contractNode)) {
-            modifications.push(makeDelete(contractNode, orig));
-          } else if (isNodeType('StructDefinition', contractNode)) {
-            const storageLocation = getStorageLocationAnnotation(contractNode);
-            if (storageLocation !== undefined) {
-              const structName = contractNode.name;
-              const variableName = `$${structName}_${(Math.random() * 1e6).toFixed(0)}`;
-              const insertText = ` ${structName} ${variableName};`;
-
-              modifications.push(makeInsertAfter(contractNode, insertText));
+          // Remove any calls to parent constructors from the inheritance list
+          const inherits = contractDef.baseContracts;
+          for (const inherit of inherits) {
+            if (Array.isArray(inherit.arguments)) {
+              assert(inherit.baseName.name !== undefined);
+              modifications.push(makeReplace(inherit, orig, inherit.baseName.name));
             }
           }
+
+          const contractNodes = contractDef.nodes;
+          for (const contractNode of contractNodes) {
+            switch (contractNode.nodeType) {
+              case 'ErrorDefinition':
+              case 'EventDefinition':
+              case 'FunctionDefinition':
+              case 'ModifierDefinition':
+              case 'UsingForDirective':
+              case 'VariableDeclaration': {
+                // Replace with an enum based on astId (the original name is not needed, since nothing should reference it)
+                modifications.push(makeReplace(contractNode, orig, toDummyEnumWithAstId(contractNode.id)));
+                break;
+              }
+              case 'StructDefinition': {
+                const storageLocation = getStorageLocationAnnotation(contractNode);
+                if (storageLocation !== undefined) {
+                  const structName = contractNode.name;
+                  const variableName = `$${structName}_${(Math.random() * 1e6).toFixed(0)}`;
+                  const insertText = ` ${structName} ${variableName};`;
+
+                  modifications.push(makeInsertAfter(contractNode, insertText));
+                }
+                break;
+              }
+              case 'EnumDefinition':
+              case 'UserDefinedValueTypeDefinition':
+              default: {
+                // - EnumDefinition may be used in structures with storage locations
+                // - UserDefinedValueTypeDefinition may be used in structures with storage locations
+                // - default: in case unexpected ast nodes show up
+                break;
+              }
+            }
+          }
+          break;
         }
-      } else if (isNodeType('FunctionDefinition', node) || isNodeType('VariableDeclaration', node)) {
-        if (node.documentation) {
-          modifications.push(makeDelete(node.documentation, orig));
+
+        // - UsingForDirective isn't needed, but it might have NatSpec documentation which is not included in the AST.
+        //   We convert it to a dummy enum to avoid orphaning any possible documentation.
+        // - ErrorDefinition, FunctionDefinition, and VariableDeclaration might be imported by other files, so they cannot be deleted.
+        //   However, we need to remove their values to avoid referencing other deleted nodes.
+        //   We do this by converting them to dummy enums, but avoiding duplicate names.
+        case 'UsingForDirective':
+        case 'ErrorDefinition':
+        case 'FunctionDefinition':
+        case 'VariableDeclaration': {
+          // If an identifier with the same name was not previously written, replace with a dummy enum using its name.
+          // Otherwise replace with an enum based on astId to avoid duplicate names, which can happen if there was overloading.
+          // This does not need to check all identifiers from the original contract, since the original compilation
+          // should have failed if there were conflicts in the first place.
+          if ('name' in node && !replacedIdentifiers.has(node.name)) {
+            modifications.push(makeReplace(node, orig, toDummyEnumWithName(node.name)));
+            replacedIdentifiers.add(node.name);
+          } else {
+            modifications.push(makeReplace(node, orig, toDummyEnumWithAstId(node.id)));
+          }
+          break;
         }
-        // Replace with a dummy variable of arbitrary type
-        const name = node.name;
-        const insertText = `uint256 constant ${name} = 0;`;
-        modifications.push(makeReplace(node, orig, insertText));
-      } else if (isNodeType('UsingForDirective', node)) {
-        modifications.push(makeDelete(node, orig));
+        case 'EnumDefinition':
+        case 'ImportDirective':
+        case 'PragmaDirective':
+        case 'StructDefinition':
+        case 'UserDefinedValueTypeDefinition':
+        default: {
+          // - EnumDefinition may be used in structures with storage locations
+          // - ImportDirective may import types used in structures with storage locations
+          // - PragmaDirective is necessary for compilation
+          // - StructDefinition may be used in structures with storage locations
+          // - UserDefinedValueTypeDefinition may be used in structures with storage locations
+          // - default: in case unexpected ast nodes show up (file-level events since 0.8.22)
+          break;
+        }
       }
     }
 
@@ -103,6 +142,14 @@ interface Modification {
   start: number;
   end: number;
   text?: string;
+}
+
+function toDummyEnumWithName(name: string) {
+  return `enum ${name} { dummy }`;
+}
+
+function toDummyEnumWithAstId(astId: number) {
+  return `enum $astId_${astId}_${(Math.random() * 1e6).toFixed(0)} { dummy }`;
 }
 
 function getPositions(node: Node) {